fix(#394): extract parseConnectorConfig to its own module for CI tests
The CI quality lane runs `tsx --test tests/**/*.test.ts` from the repo
root without building workspace packages first. The
`tests/cli-parse-connector-config.test.ts` file was importing from
`packages/remnic-cli/src/index.ts`, which transitively imports from
`@remnic/core`. Node resolved `@remnic/core` to its `dist/index.js`
export, which does not exist in CI before the build step, producing:

  ERR_MODULE_NOT_FOUND: Cannot find module
  '.../packages/remnic-cli/node_modules/@remnic/core/dist/index.js'

Fix by extracting `parseConnectorConfig` into a standalone, dependency-
free helper at `packages/remnic-cli/src/parse-connector-config.ts`. The
CLI entry re-exports it for backwards compatibility; the test imports
directly from the helper, mirroring the existing
`service-candidates.ts` pattern in `tests/remnic-cli-service-candidates.test.ts`.
